support for Max Payne 2 Static Meshes from LDB
