+++ max_payne_sdk/ldb_common/max_ldb_interface.py
@@ -1,0 +1,2 @@
+class MaxLDBInterface:
+    pass
